Clean up incremental compilation tests
